refactor(eval): split god command file into tree/actions/helpers
eval.go had grown to ~970 lines interleaving cobra wiring, run behavior,
and utility helpers across 13 subcommands. Split along its natural seams:
  - eval.go         cobra command tree + evalRunFlags (the command surface)
  - eval_actions.go runEval* behavior + printBenchmarkSummary + TUI data
  - eval_helpers.go skill/workspace resolution, scaffolding, small utils

Pure code move within package main; no behavior change. Verified the full
eval subcommand tree, JSON output, and a mock eval run end-to-end.
